Extract shell snapshot types
